+++ ConsoleApplication1/ConsoleApplication1/Tree.cpp
@@ -1,7 +1,7 @@
 #include "Tree.hpp"
 
 
-Character* Tree::Clone(unsigned int x, unsigned int y) {
+Character* Tree::Clone(float x, float y) {
 	Tree* t = new Tree(*this);
 	t->SetPosition(x, y);
 	t->_ID = t->lastID++;
