+++ ConsoleApplication1/ConsoleApplication1/Tree.hpp
@@ -11,12 +11,12 @@ public:
 
 
 	std::string Falar() override { return "Eu sou uma árvore. Eu não falo, idiota."; }
-	void Andar(int deltax, int deltay) override {}
+	void Andar(float deltax, float deltay) override {}
 	void Lutar() override {}
 	void DoEvents(void* charmanager, Map* map) override {}
 	void Reviver() override {}
 
 
-	Character* Clone(unsigned int x, unsigned int y) override;
+	Character* Clone(float x, float y) override;
 
 };
